LayoutTests/accessibility/mac/html-section-elements-expected.txt

header
footer
section
article
nav
aside
Should be IGNORED because it's inside an article, so it should not be a landmark
Should be IGNORED because it's inside an article, so it should not be a landmark
Should be IGNORED because it's inside a section, so it should not be a landmark
Should be IGNORED because it's inside a section, so it should not be a landmark
This tests that the HTML5 section elements map to the correct AX roles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).subrole is 'AXSubrole: AXLandmarkBanner'
PASS body.childAtIndex(1).subrole is 'AXSubrole: AXLandmarkContentInfo'
PASS body.childAtIndex(2).subrole is 'AXSubrole: AXDocumentRegion'
PASS body.childAtIndex(3).subrole is 'AXSubrole: AXDocumentArticle'
PASS body.childAtIndex(4).subrole is 'AXSubrole: AXLandmarkNavigation'
PASS body.childAtIndex(5).subrole is 'AXSubrole: AXLandmarkComplementary'
PASS body.childAtIndex(6).childAtIndex(0).subrole is 'AXSubrole: '
PASS body.childAtIndex(7).childAtIndex(0).subrole is 'AXSubrole: '
PASS successfullyParsed is true

TEST COMPLETE
